chore: update data types to mc priorty queue
