+++ src/clients/jobManagerClient.ts
@@ -1,9 +1,10 @@
 import { Logger } from '@map-colonies/js-logger';
+import { IJobResponse, IUpdateJobBody } from '@map-colonies/mc-priority-queue';
 import { HttpClient, IHttpRetryConfig } from '@map-colonies/mc-utils';
 import config from 'config';
 import { injectable, inject } from 'tsyringe';
 import { Services } from '../common/constants';
-import { IJobResponse, IUpdateJobBody } from '../common/interfaces/jobs';
+import { IJobParameters, ITaskParameters } from '../common/interfaces/interfaces';
 
 @injectable()
 export class JobManagerClient extends HttpClient {
@@ -11,13 +12,13 @@ export class JobManagerClient extends HttpClient {
     super(logger, config.get<string>('queue.jobManagerBaseUrl'), 'JobManager', config.get<IHttpRetryConfig>('httpRetry'));
   }
 
-  public async getJob(jobId: string): Promise<IJobResponse> {
+  public async getJob(jobId: string): Promise<IJobResponse<IJobParameters, ITaskParameters>> {
     this.logger.debug(`Retrieving job ${jobId}`);
-    const job = await this.get<IJobResponse>(`/jobs/${jobId}`);
+    const job = await this.get<IJobResponse<IJobParameters, ITaskParameters>>(`/jobs/${jobId}`);
     return job;
   }
 
-  public async updateJob(jobId: string, payload: IUpdateJobBody): Promise<void> {
+  public async updateJob(jobId: string, payload: IUpdateJobBody<IJobParameters>): Promise<void> {
     this.logger.debug(`Updating job ${jobId} with payload ${JSON.stringify(payload)}`);
     const updateJobUrl = `/jobs/${jobId}`;
     await this.put(updateJobUrl, payload);
