+++ src/common/interfaces/interfaces.ts
@@ -1,4 +1,4 @@
-import { MultiPolygon, Polygon } from '@turf/helpers';
+import { Polygon, MultiPolygon } from '@turf/helpers';
 import { BBox2d } from '@turf/helpers/dist/js/lib/geojson';
 
 export interface IConfig {
@@ -48,4 +48,31 @@ export interface ICallbackResponse {
   requestId: string;
   success: boolean;
   errorReason?: string;
+}
+export interface IJobParameters {
+  dbId: string;
+  targetResolution: number;
+  crs: string;
+  callbackURL: string[];
+  bbox: BBox2d;
+  priority: number;
+  packageName: string;
+  footprint: Polygon | MultiPolygon;
+  version: string;
+  cswProductId: string;
+  tilesPath: string;
+  callbackParams?: ICallbackResponse;
+}
+
+export interface ITaskParameters {
+  callbackURL: string[];
+  bbox: BBox2d | true;
+  dbId: string;
+  footprint: Polygon | MultiPolygon;
+  tilesPath: string;
+  zoomLevel: number;
+  packageName: string;
+  productType: string;
+  crs: string;
+  priority: number;
 }
